add policyDTO, key management and include on GetAll

+++ AllPolicyInsurance/Models/Address.cs
@@ -5,7 +5,7 @@ namespace AllPolicyInsurance.Models
     public class Address
     {
         [Key]
-        public int AddressId { get; set; }
+        public byte AddressId { get; set; }
         public string Street { get; set; }
         public string City { get; set; }
         public string State { get; set; }
