+++ AllPolicyInsurance/Models/Vehicle.cs
@@ -5,7 +5,7 @@ namespace AllPolicyInsurance.Models
     public class Vehicle
     {
         [Key]
-        public int VehicleId { get; set; }
+        public byte VehicleId { get; set; }
         public int Year { get; set; }
         public string Model { get; set; }
         public string Make { get; set; }
